Solving problem of update_server getting id 0
The update server of accredited translation could get the ID 0 causing problems.

This new SQL sets its id to 10000 and the auto increment to 10001. That ensures not to override existing update servers in the #__update_sites table.

And also leave plenty of space to Joomla Core to add specific update servers in between 2 and 9999 update_servers